Update Logarithm.cpp
Finished most of simplify method for int base and int operand. Need some help with how to deal with expression as an operand/base but once i get a few steps into it should be easy to recreate the simplify method for the other 3 cases (int or expression base and int or expression operand combinations).